# backend-go/internal/api/api.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.1 DO NOT EDIT.
package api

import (
	"fmt"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

const (
	BearerAuthScopes = "bearerAuth.Scopes"
)

// ContactRequest defines model for ContactRequest.
type ContactRequest struct {
	Email   openapi_types.Email `json:"email"`
	Message string              `json:"message"`
	Name    string              `json:"name"`
	Subject *string             `json:"subject,omitempty"`
}

// Experience defines model for Experience.
type Experience struct {
	UnderscoreId string    `json:"_id"`
	Company      string    `json:"company"`
	Description  *[]string `json:"description,omitempty"`
	End          *string   `json:"end,omitempty"`
	Location     *string   `json:"location,omitempty"`
	Order        *int      `json:"order,omitempty"`
	Role         string    `json:"role"`
	Start        string    `json:"start"`
	Tags         *[]string `json:"tags,omitempty"`
	Url          *string   `json:"url,omitempty"`
}

// LoginRequest defines model for LoginRequest.
type LoginRequest struct {
	Email    openapi_types.Email `json:"email"`
	Password string              `json:"password"`
}

// LoginResponse defines model for LoginResponse.
type LoginResponse struct {
	Token string `json:"token"`
}

// MessageResponse defines model for MessageResponse.
type MessageResponse struct {
	Message string `json:"message"`
}

// Project defines model for Project.
type Project struct {
	UnderscoreId string    `json:"_id"`
	Contributors *string   `json:"contributors,omitempty"`
	Details      *[]string `json:"details,omitempty"`
	Image        *string   `json:"image,omitempty"`
	Links        struct {
		Demo *string `json:"demo,omitempty"`
		Repo *string `json:"repo,omitempty"`
	} `json:"links"`
	Order   *int     `json:"order,omitempty"`
	Slug    string   `json:"slug"`
	Stack   []string `json:"stack"`
	Summary string   `json:"summary"`
	Title   string   `json:"title"`
}

// ReorderItem defines model for ReorderItem.
type ReorderItem struct {
	Id    string `json:"id"`
	Order int    `json:"order"`
}

// SkillCategory defines model for SkillCategory.
type SkillCategory struct {
	UnderscoreId string       `json:"_id"`
	Items        *[]SkillItem `json:"items,omitempty"`
	Order        *int         `json:"order,omitempty"`
	Title        string       `json:"title"`
}

// SkillItem defines model for SkillItem.
type SkillItem struct {
	Description *string `json:"description,omitempty"`
	Icon        *string `json:"icon,omitempty"`
	Level       string  `json:"level"`
	Name        string  `json:"name"`
}

// IdParam defines model for IdParam.
type IdParam = string

// ReorderExperienceJSONBody defines parameters for ReorderExperience.
type ReorderExperienceJSONBody = []ReorderItem

// ReorderProjectsJSONBody defines parameters for ReorderProjects.
type ReorderProjectsJSONBody = []ReorderItem

// ReorderSkillsJSONBody defines parameters for ReorderSkills.
type ReorderSkillsJSONBody = []ReorderItem

// PostLoginJSONRequestBody defines body for PostLogin for application/json ContentType.
type PostLoginJSONRequestBody = LoginRequest

// PostContactJSONRequestBody defines body for PostContact for application/json ContentType.
type PostContactJSONRequestBody = ContactRequest

// CreateExperienceJSONRequestBody defines body for CreateExperience for application/json ContentType.
type CreateExperienceJSONRequestBody = Experience

// ReorderExperienceJSONRequestBody defines body for ReorderExperience for application/json ContentType.
type ReorderExperienceJSONRequestBody = ReorderExperienceJSONBody

// UpdateExperienceJSONRequestBody defines body for UpdateExperience for application/json ContentType.
type UpdateExperienceJSONRequestBody = Experience

// CreateProjectJSONRequestBody defines body for CreateProject for application/json ContentType.
type CreateProjectJSONRequestBody = Project

// ReorderProjectsJSONRequestBody defines body for ReorderProjects for application/json ContentType.
type ReorderProjectsJSONRequestBody = ReorderProjectsJSONBody

// UpdateProjectJSONRequestBody defines body for UpdateProject for application/json ContentType.
type UpdateProjectJSONRequestBody = Project

// CreateSkillCategoryJSONRequestBody defines body for CreateSkillCategory for application/json ContentType.
type CreateSkillCategoryJSONRequestBody = SkillCategory

// ReorderSkillsJSONRequestBody defines body for ReorderSkills for application/json ContentType.
type ReorderSkillsJSONRequestBody = ReorderSkillsJSONBody

// UpdateSkillCategoryJSONRequestBody defines body for UpdateSkillCategory for application/json ContentType.
type UpdateSkillCategoryJSONRequestBody = SkillCategory

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Admin login
	// (POST /auth/login)
	PostLogin(c *gin.Context)
	// List contact messages (admin)
	// (GET /contact)
	GetContactMessages(c *gin.Context)
	// Submit contact form
	// (POST /contact)
	PostContact(c *gin.Context)
	// Get work experience list
	// (GET /experience)
	GetExperience(c *gin.Context)
	// Create a new experience entry
	// (POST /experience)
	CreateExperience(c *gin.Context)
	// Reorder experience entries
	// (PUT /experience/reorder)
	ReorderExperience(c *gin.Context)
	// Delete an experience entry
	// (DELETE /experience/{id})
	DeleteExperience(c *gin.Context, id IdParam)
	// Update an experience entry
	// (PUT /experience/{id})
	UpdateExperience(c *gin.Context, id IdParam)
	// Get all portfolio projects
	// (GET /projects)
	GetProjects(c *gin.Context)
	// Create a new project
	// (POST /projects)
	CreateProject(c *gin.Context)
	// Reorder projects
	// (PUT /projects/reorder)
	ReorderProjects(c *gin.Context)
	// Delete a project
	// (DELETE /projects/{id})
	DeleteProject(c *gin.Context, id IdParam)
	// Update a project
	// (PUT /projects/{id})
	UpdateProject(c *gin.Context, id IdParam)
	// Get skills categorized
	// (GET /skills)
	GetSkills(c *gin.Context)
	// Create a new skill category
	// (POST /skills)
	CreateSkillCategory(c *gin.Context)
	// Reorder skill categories
	// (PUT /skills/reorder)
	ReorderSkills(c *gin.Context)
	// Delete a skill category
	// (DELETE /skills/{id})
	DeleteSkillCategory(c *gin.Context, id IdParam)
	// Update a skill category
	// (PUT /skills/{id})
	UpdateSkillCategory(c *gin.Context, id IdParam)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandler       func(*gin.Context, error, int)
}

type MiddlewareFunc func(c *gin.Context)

// PostLogin operation middleware
func (siw *ServerInterfaceWrapper) PostLogin(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostLogin(c)
}

// GetContactMessages operation middleware
func (siw *ServerInterfaceWrapper) GetContactMessages(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetContactMessages(c)
}

// PostContact operation middleware
func (siw *ServerInterfaceWrapper) PostContact(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.PostContact(c)
}

// GetExperience operation middleware
func (siw *ServerInterfaceWrapper) GetExperience(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetExperience(c)
}

// CreateExperience operation middleware
func (siw *ServerInterfaceWrapper) CreateExperience(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateExperience(c)
}

// ReorderExperience operation middleware
func (siw *ServerInterfaceWrapper) ReorderExperience(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ReorderExperience(c)
}

// DeleteExperience operation middleware
func (siw *ServerInterfaceWrapper) DeleteExperience(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IdParam

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteExperience(c, id)
}

// UpdateExperience operation middleware
func (siw *ServerInterfaceWrapper) UpdateExperience(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IdParam

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateExperience(c, id)
}

// GetProjects operation middleware
func (siw *ServerInterfaceWrapper) GetProjects(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetProjects(c)
}

// CreateProject operation middleware
func (siw *ServerInterfaceWrapper) CreateProject(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateProject(c)
}

// ReorderProjects operation middleware
func (siw *ServerInterfaceWrapper) ReorderProjects(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ReorderProjects(c)
}

// DeleteProject operation middleware
func (siw *ServerInterfaceWrapper) DeleteProject(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IdParam

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteProject(c, id)
}

// UpdateProject operation middleware
func (siw *ServerInterfaceWrapper) UpdateProject(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IdParam

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateProject(c, id)
}

// GetSkills operation middleware
func (siw *ServerInterfaceWrapper) GetSkills(c *gin.Context) {

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.GetSkills(c)
}

// CreateSkillCategory operation middleware
func (siw *ServerInterfaceWrapper) CreateSkillCategory(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.CreateSkillCategory(c)
}

// ReorderSkills operation middleware
func (siw *ServerInterfaceWrapper) ReorderSkills(c *gin.Context) {

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.ReorderSkills(c)
}

// DeleteSkillCategory operation middleware
func (siw *ServerInterfaceWrapper) DeleteSkillCategory(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IdParam

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.DeleteSkillCategory(c, id)
}

// UpdateSkillCategory operation middleware
func (siw *ServerInterfaceWrapper) UpdateSkillCategory(c *gin.Context) {

	var err error

	// ------------- Path parameter "id" -------------
	var id IdParam

	err = runtime.BindStyledParameterWithOptions("simple", "id", c.Param("id"), &id, runtime.BindStyledParameterOptions{Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandler(c, fmt.Errorf("Invalid format for parameter id: %w", err), http.StatusBadRequest)
		return
	}

	c.Set(BearerAuthScopes, []string{})

	for _, middleware := range siw.HandlerMiddlewares {
		middleware(c)
		if c.IsAborted() {
			return
		}
	}

	siw.Handler.UpdateSkillCategory(c, id)
}

// GinServerOptions provides options for the Gin server.
type GinServerOptions struct {
	BaseURL      string
	Middlewares  []MiddlewareFunc
	ErrorHandler func(*gin.Context, error, int)
}

// RegisterHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterHandlers(router gin.IRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, GinServerOptions{})
}

// RegisterHandlersWithOptions creates http.Handler with additional options
func RegisterHandlersWithOptions(router gin.IRouter, si ServerInterface, options GinServerOptions) {
	errorHandler := options.ErrorHandler
	if errorHandler == nil {
		errorHandler = func(c *gin.Context, err error, statusCode int) {
			c.JSON(statusCode, gin.H{"msg": err.Error()})
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandler:       errorHandler,
	}

	router.POST(options.BaseURL+"/auth/login", wrapper.PostLogin)
	router.GET(options.BaseURL+"/contact", wrapper.GetContactMessages)
	router.POST(options.BaseURL+"/contact", wrapper.PostContact)
	router.GET(options.BaseURL+"/experience", wrapper.GetExperience)
	router.POST(options.BaseURL+"/experience", wrapper.CreateExperience)
	router.PUT(options.BaseURL+"/experience/reorder", wrapper.ReorderExperience)
	router.DELETE(options.BaseURL+"/experience/:id", wrapper.DeleteExperience)
	router.PUT(options.BaseURL+"/experience/:id", wrapper.UpdateExperience)
	router.GET(options.BaseURL+"/projects", wrapper.GetProjects)
	router.POST(options.BaseURL+"/projects", wrapper.CreateProject)
	router.PUT(options.BaseURL+"/projects/reorder", wrapper.ReorderProjects)
	router.DELETE(options.BaseURL+"/projects/:id", wrapper.DeleteProject)
	router.PUT(options.BaseURL+"/projects/:id", wrapper.UpdateProject)
	router.GET(options.BaseURL+"/skills", wrapper.GetSkills)
	router.POST(options.BaseURL+"/skills", wrapper.CreateSkillCategory)
	router.PUT(options.BaseURL+"/skills/reorder", wrapper.ReorderSkills)
	router.DELETE(options.BaseURL+"/skills/:id", wrapper.DeleteSkillCategory)
	router.PUT(options.BaseURL+"/skills/:id", wrapper.UpdateSkillCategory)
}